% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.sd_package}
\alias{as.sd_package}
\title{Return information about a package}
\usage{
as.sd_package(pkg = ".", site_path = NULL, examples = NULL)
}
\arguments{
  \item{package}{name of package, as character vector}

  \item{base_path}{root directory in which to create
  documentation. The default, \code{NULL}, first looks at
  the value of \code{site_path} set in \file{DESCRIPTION},
  and if not found uses \code{inst/web}.}

  \item{examples}{include examples or not? The default,
  \code{NULL}, first looks at the value of \code{examples}
  set in \file{DESCRIPTION}, and if not found uses
  \code{TRUE}.}
}
\value{
A named list of useful metadata about a package
}
\description{
Return information about a package
}
\keyword{internal}

