% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-search-docs.R
\name{build_docsearch_json}
\alias{build_docsearch_json}
\title{Build Docsearch JSON}
\usage{
build_docsearch_json(pkg = ".")
}
\description{
Create a Docsearch search configuration in JSON format and save it as
\code{docsearch.json} to be used as a suggested starting point when setting up a
Docsearch crawler.
}
