% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-search-docs.R
\name{build_search}
\alias{build_search}
\title{Build search index}
\usage{
build_search(pkg = ".", override = list())
}
\arguments{
\item{pkg}{Path to package.}

\item{override}{}
}
\description{
Build a JSON file encompassing all HTML pages, for use by the search script.
}
\section{YAML config}{

You can exclude some paths from the search index.\if{html}{\out{<div class="yaml">}}\preformatted{search:
  exclude: ['news/index.html']
}\if{html}{\out{</div>}}
}

\section{Local testing}{

Search won't work if you simply use pkgdown preview of the static files.
You can use \code{servr::httw("docs")} instead.
}

