% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-search-docs.R
\name{build_search_docs}
\alias{build_search_docs}
\title{Build documents for search infrastructure}
\usage{
build_search_docs(pkg = ".")
}
\description{
These documents are built under \code{docs} as part of the Docsearch
infrastructure:
}
\details{
\itemize{
\item \code{docsearch.json}
}
}
