% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-search-index.R
\name{build_search_index}
\alias{build_search_index}
\title{Build search index}
\usage{
build_search_index(pkg = ".", path = "docs", depth = 0L,
  index_rd = TRUE, index_vignette = TRUE, vignette_path = "docs/articles")
}
\arguments{
\item{pkg}{Path to source package. If R working directory is not
set to the source directory, then pkg must be a fully qualified
path to the source directory (not a relative path).}

\item{path}{Output path. Relative paths are taken relative to the
\code{pkg} directory.}

\item{depth}{Depth of path relative to root of documentation.  Used
to adjust relative links in the navbar.}

\item{index_rd}{include Rd files in search index}

\item{index_vignette}{include vignettes in search index}

\item{vignette_path}{path to rendered vignettes}
}
\description{
By default pkgdown will create a site search index compiled from Rd file
descriptions and vignette content.
}
\details{
Data for the index is saved in JSON format in \code{site-index.json}.
}
