% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.r
\name{build_site_rstudio}
\alias{build_site_rstudio}
\title{Document, build, and preview site in RStudio}
\usage{
build_site_rstudio(pkg = ".")
}
\description{
\code{build_site_rstudio()} is a wrapper around these three methods:
\itemize{
\item \code{\link[devtools:document]{devtools::document()}}
\item \code{\link[pkgdown:build_site]{pkgdown::build_site()}}
\item \code{\link[pkgdown:preview_site]{pkgdown::preview_site()}}
}
}
