\name{build_topics}
\alias{build_topics}
\title{Generate all topic pages for a package.}
\usage{
  build_topics(package)
}
\arguments{
  \item{package_info}{A list containing information about
  the package, as generated by \code{\link{package_info}}}

  \item{package}{name of package (as character string) to
  create documentation for}
}
\description{
  Generate all topic pages for a package.
}

