% Generated by roxygen2 (4.0.0): do not edit by hand
\name{build_web}
\alias{build_package}
\alias{build_web}
\alias{staticdocs-package}
\title{Build complete static documentation for a package.}
\usage{
build_web(package = ".", base_path = "inst/web", examples = NULL,
  templates_path = NULL, launch = interactive())
}
\arguments{
  \item{package}{path to source version of package.  See
  \code{\link[devtools]{as.package}} for details on how
  paths and package names are resolved.}

  \item{base_path}{root directory in which to create
  documentation}

  \item{examples}{include examples or not?  Examples are
  particularly slow to render because all code must be run,
  so turning them off makes it easier to tweak templates
  etc.}

  \item{templates_path}{a specific directory path to use
  when searching for rendering templates, in addition to
  the default locations of packagedir/inst/staticdocs,
  packagedir/staticdocs, and the staticdocs package's
  included templates directory.}

  \item{launch}{If \code{TRUE}, will open freshly generated
  site in web browser.}
}
\description{
Currently, knitr builds documentation for:
}
\details{
\itemize{
  \item Rd files.  Files
  \item Demos. Must be listed in \file{demos/00index}.
  \item Vignettes.
}
}

