% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_index.R
\name{check_index_missing}
\alias{check_index_missing}
\title{Check Articles and Reference Index}
\usage{
check_index_missing(pkg = ".", check_articles = TRUE, check_reference = TRUE)
}
\arguments{
\item{pkg}{Path to package.}

\item{check_articles}{Whether to check articles index}

\item{check_reference}{Whether to check reference index.}
}
\value{
If no errors are found returns \code{TRUE} otherwise a warning.
}
\description{
Check Articles and Reference Index
}
