% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_pkgdown}
\alias{check_pkgdown}
\title{Check \verb{_pkgdown.yml}}
\usage{
check_pkgdown(pkg = ".")
}
\arguments{
\item{pkg}{Path to package.}
}
\description{
Check that your \verb{_pkgdown.yml} is valid without building the whole
site.
}
