% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy-site.R
\name{ci_commit_sha}
\alias{ci_commit_sha}
\title{Commit SHA Value}
\usage{
ci_commit_sha()
}
\description{
Attempts to obtain the commit SHA identifier the CI platform is using to
build the package and refresh the website.
}
\details{
Depending on the CI environment used, the underlying commit SHA variable
may be named differently. Below are the supported CI variables for retrieving
a commit ID.
\itemize{
\item Travis-CI: \href{https://docs.travis-ci.com/user/environment-variables/#default-environment-variables}{\code{TRAVIS_COMMIT}}
\item GitHub Actions: \href{https://help.github.com/en/actions/configuring-and-managing-workflows/using-environment-variables#default-environment-variables}{\code{GITHUB_SHA}}
}
}
\keyword{internal}
