% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy-site.R
\name{deploy_local}
\alias{deploy_local}
\title{Build and deploy a site locally}
\usage{
deploy_local(pkg = ".", commit_message = construct_commit_message(pkg), ...)
}
\arguments{
\item{pkg}{Path to package.}

\item{commit_message}{The commit message to be used for the commit.}

\item{...}{Additional arguments passed to \code{\link[=build_site]{build_site()}}.}
}
\description{
Build and deploy a site locally
}
