% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_emep.R
\name{ef_emep}
\alias{ef_emep}
\alias{ef_emep_exh}
\alias{ef_emep_exh_pm2.5}
\alias{ef_emep_exh_pm10}
\title{EMEP Emissions Factors}
\usage{
ef_emep_exh(
  veh.spd = NULL,
  veh.wt = NULL,
  em.type = "pm2.5",
  route.def = NULL,
  em.source = "exhaust",
  euro.class = NULL,
  ...
)

ef_emep_exh_pm2.5(...)

ef_emep_exh_pm10(...)
}
\arguments{
\item{veh.spd}{(numeric, required) vehicle speed (in km/hr).}

\item{veh.wt}{(numeric, required) vehicle wt (in kg).}

\item{em.type}{(character) type of emissions to predict, just PM at the moment.}

\item{route.def}{(character) route description or identifier.}

\item{em.source}{(character) emission source, just exhaust at the moment.}

\item{euro.class}{(numeric at moment) the emission control classification,
currently as number 0 (pre euro) to euro 6.}

\item{...}{other arguments, often passed on.}
}
\value{
These functions build data.frames of vehicle speed based emission
factors.
}
\description{
Functions to estimate vehicle speed-based emission factors
based on EMEP handbook methods.

IMPORTANT: THESE CURRENTLY ONLY WORK FOR EURO VI BUSES...
}
\note{
These may be moving to vein at some point...
}
\references{
These functions are based on methods developed and reported by:

\link{handbook}
}
