% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs.main.R
\name{embrs.main}
\alias{embrs.main}
\alias{build_inventory}
\title{embrs}
\usage{
build_inventory(x, ...)
}
\arguments{
\item{x}{(required embrs object) the \strong{embrs} model to analyse.}
}
\value{
These functions return different \strong{embrs} outputs.
}
\description{
Main functions for embrs emission model anslysis.
}
\note{
These may be moving to vein at some point...
}
\references{
add embrs and Beddows references...
}
