% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{fig_settings}
\alias{fig_settings}
\title{You will generally not need to use this unless you are handling
custom plot output.}
\usage{
fig_settings()
}
\value{
A list containing the entries from the \code{figures} field
in \verb{_pkgdown.yaml} (see \link{build_reference}), with default values added.
Computed \code{width} and \code{height} values (in pixels) are also included.
}
\description{
You will generally not need to use this unless you are handling
custom plot output.
}
