% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_align.R
\name{fit_align}
\alias{fit_align}
\alias{fit_align.default}
\title{fit_align}
\usage{
\method{fit_align}{default}(
  x,
  y = NULL,
  by = NULL,
  fun = function(par, x) par[1] + (par[2] * x),
  lower = c(-60, 0.3),
  upper = c(60, 3),
  ...
)
}
\arguments{
\item{x}{A first time-series \code{vector} or a \code{data.frame}
containing time-series to use a reference when warp-fitting
\code{y}.}

\item{y}{A second time-series \code{vector} or a \code{data.frame}
containing time-series to warp-fit with \code{x}.}

\item{by}{If \code{x} or \code{y} are \code{data.frames}, the names
of time-series to warp-fit.}

\item{fun}{A \code{function} describing the warp-fitting model
to be applied.}

\item{lower}{The lowest values for \code{fun}, one per parameter.}

\item{upper}{The highest values for \code{fun}, one per parameter.}

\item{...}{Other arguments, typically applied by generic
\code{alignment} methods or ignored.}
}
\value{
By default, \code{fit_align} returns \code{x} and \code{y}
as an aligned \code{data.frame}. It also provides a lag correlation
profile (plot) and alignment report. The extra \code{function},
\code{output}, can be used to modify this behavior (see also
\code{\link{alignment}}).
}
\description{
Time warp a data-series using a time-offsetting model.
}
\author{
Karl Ropkins
}
