% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnote.cite.R
\name{footnote.cite}
\alias{footnote.cite}
\title{footnote.cite}
\usage{
footnote.cite(x = "", linebreak = FALSE)
}
\arguments{
\item{x}{The name of the package to be cited. The default cites R itself.}

\item{linebreak}{force linebreaks. See Note}
}
\description{
footnote.cite attempts to tidy citations so they can be used in document footnotes...
}
\note{
Linebreaking would typically be left to the code creating the footnote. However,
if you are using \code{footnote.cite()} in other applications \code{linebreak=TRUE} may help
}
\examples{
#to do something
}
\keyword{methods}
