% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{index}
\alias{index}
\alias{test-index}
\title{Index}
\description{
Index
}
\seealso{
Other tests:
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-long-lines}},
\code{\link{test-math-examples}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-sexpr-title}},
\code{\link{test-tables}},
\code{\link{test-verbatim}}
}
\concept{tests}
\keyword{internal}
