% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R, R/build-favicons.R
\name{init_site}
\alias{init_site}
\alias{build_favicons}
\title{Initialise site infrastructure}
\usage{
init_site(pkg = ".")

build_favicons(pkg = ".", overwrite = FALSE)
}
\arguments{
\item{pkg}{Path to package.}

\item{overwrite}{If \code{TRUE}, re-create favicons from package logo.}
}
\description{
\code{init_site()}:
\itemize{
\item creates the output directory (\verb{docs/}),
\item generates a machine readable description of the site, used for autolinking,
\item copies CSS/JS assets and extra files, and
\item runs \code{build_favicons()}, if needed.
}

\code{build_favicons()}:
\itemize{
\item creates favicons from package logo.
\item auto-detects the location of your package logo (with the name \code{logo.svg}
(recommended format) or \code{logo.png})
\item runs it through the \url{https://realfavicongenerator.net} API to build a complete
set of favicons with different sizes, as needed for modern web usage.
}

See \code{vignette("customise")} for the various ways you can customise the
display of your site.
}
\details{
You only need to run \code{build_favicons()} once. The favicon set will be stored in
\code{pkgdown/favicon} and copied by \code{init_site()} to the relevant location when
the website is rebuilt.
}
\section{Build-ignored files}{
We recommend using \code{\link[usethis:use_pkgdown]{usethis::use_pkgdown()}} to build-ignore \verb{docs/} and
\verb{_pkgdown.yml}. If use another directory, or create the site manually,
you'll need to add them to \code{.Rbuildignore} yourself. A \code{NOTE} about
an unexpected file during \verb{R CMD CHECK} is an indication you have not
correctly ignored these files.
}

