% Generated by roxygen2 (4.0.0): do not edit by hand
\name{package_info}
\alias{package_info}
\title{Return information about a package}
\usage{
package_info(package = ".", base_path = "staticdocs", examples = NULL,
  templates_path = NULL)
}
\arguments{
  \item{package}{name of package, as character vector}
}
\value{
A named list of useful metadata about a package
}
\description{
Return information about a package
}
\keyword{internal}

