% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rd.r
\name{parse_rd}
\alias{parse_rd}
\title{Parse an rd file in to staticdocs format.}
\usage{
parse_rd(topic, package)
}
\arguments{
\item{topic}{topic name, as character vector}

\item{package}{package name, as character vector}
}
\description{
Rd files are pretty printed with structural elements coloured blue, and
leaves are given a short prefix: \code{\"} = text, \code{\'} = verbatim, and
\code{>} = R code.
}
\examples{
parse_rd("whisker.render", "whisker")
}

