% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.r
\name{parse_topic}
\alias{parse_topic}
\title{Parse a help topic in to staticdocs format.}
\usage{
parse_topic(topic, package)
}
\arguments{
\item{topic}{topic name, as character vector}

\item{package}{package name, as character vector}
}
\description{
Rd files are pretty printed with structural elements coloured blue, and
leaves are given a short prefix: \code{\"} = text, \code{\'} = verbatim, and
\code{>} = R code.
}
\examples{
parse_topic("whisker.render", "whisker")
}

