% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{pkgdown_sitrep}
\alias{pkgdown_sitrep}
\title{Report package pkgdown situation}
\usage{
pkgdown_sitrep(pkg = ".")
}
\arguments{
\item{pkg}{Path to package.}
}
\description{
\code{pkgdown_sitrep()} reports
\itemize{
\item If there is an \code{url} field in the pkgdown configuration;
\item If that pkgdown website URL is stored in the DESCRIPTION file.
}
}
