% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_page}
\alias{preview_page}
\title{Preview a local pkgdown page in the browser}
\usage{
preview_page(path, pkg = ".")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[=preview_site]{preview_site()}} instead.
}
\keyword{internal}
