% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.R
\name{pulledpork}
\alias{pulledpork}
\title{Pulled pork}
\description{
Pulled pork
}
\examples{
1 + 2
plot(1, 1, axes = FALSE)

\dontrun{
1 + 3
}

2 + 3
plot(1:2, 1:2, axes = FALSE)

\donttest{
1 + 3
}

\dontshow{
5 + 6
}
}
\keyword{internal}
