% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd2html}
\alias{rd2html}
\title{Translate an Rd string to its HTML output}
\usage{
rd2html(x, fragment = TRUE, ...)
}
\arguments{
\item{x}{Rd string. Backslashes must be double-escaped ("\\\\").}

\item{fragment}{logical indicating whether this represents a complete Rd file}

\item{...}{additional arguments for as_html}
}
\description{
Translate an Rd string to its HTML output
}
\examples{
rd2html("a\n\%b\nc")

rd2html("a & b")

rd2html("\\\\strong{\\\\emph{x}}")

}
