% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{rel_path}
\alias{rel_path}
\title{Compute relative path}
\usage{
rel_path(path, base = ".", windows = on_windows())
}
\arguments{
\item{path}{Relative path}

\item{base}{Base path}

\item{windows}{Whether the operating system is Windows. Default value is to
check the user's system information.}
}
\description{
Compute relative path
}
\examples{
rel_path("a/b", base = "here")
rel_path("/a/b", base = "here")
}
