% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.r
\name{render_page}
\alias{render_page}
\title{Render complete page.}
\usage{
render_page(package, name, data, path = "", depth = 0L)
}
\arguments{
\item{package}{Path to package to document.}

\item{name}{Name of the template (e.g. index, demo, topic)}

\item{data}{Data for the template. Package metadata is always automatically
added to this list under key \code{package}.}

\item{path}{Location to create file. If \code{""} (the default),
prints to standard out.}

\item{depth}{Depth of path relative to base directory. Used to
adjust links in navbar, and to provide template variable \code{root_path}.}
}
\description{
Render complete page.
}

