% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/icons.r
\name{sd_icon}
\alias{sd_icon}
\title{Define the icon for a function.}
\usage{
sd_icon(expr = NULL, func = NULL, inherits = NULL)
}
\arguments{
\item{expr,func}{Either a bare expression or a function with no arguments
that uses grid to create a icon that represents the function.}

\item{inherits}{Alternatively, use an existing icon specified by a
function name}
}
\description{
Define the icon for a function.
}

