\name{sd_item}
\alias{sd_item}
\title{Define an item in a section of the index.}
\usage{
  sd_item(name, title = NULL)
}
\arguments{
  \item{name}{name of the function}

  \item{title}{override the default title extracted from
  the corresponding Rd file}
}
\description{
  Define an item in a section of the index.
}

