\name{sd_section}
\alias{sd_section}
\title{Define a section for the index page}
\usage{
  sd_section(name, description, elements)
}
\arguments{
  \item{name}{Name of the section. Used as title.}

  \item{description}{Paragraph description of the section.
  May use markdown.}

  \item{elements}{Either a list containing either strings
  giving function names, or if you want to override
  defaults from the rd file,}
}
\description{
  Define a section for the index page
}

