% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{template_navbar}
\alias{template_navbar}
\title{Generate YAML for default navbar}
\usage{
template_navbar(path = ".")
}
\description{
Use this function if you want to customise the display of the default
navbar.
}

