% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-long-lines}
\alias{test-long-lines}
\title{Test case: long-lines}
\description{
The example results should have the copy button correctly placed when
scrollings
}
\examples{
pkgdown:::ruler()

cat(rep("x ", 100), sep = "")
cat(rep("xy", 100), sep = "")
cat(rep("x ", 100), sep = "")
cat(rep("xy", 100), sep = "")
}
\seealso{
Other tests:
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-math-examples}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-sexpr-title}},
\code{\link{test-tables}},
\code{\link{test-verbatim}}
}
\concept{tests}
\keyword{internal}
