% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-math-examples}
\alias{test-math-examples}
\title{Test case: math rendering in examples}
\arguments{
\item{x}{\eqn{f(x) > 0}: inline equation}
}
\description{
Test case: math rendering in examples
}
\details{
Display equation:

\deqn{y = \alpha + \beta X + \varepsilon}

Multi-line equation (correctly rendered by katex only):

\deqn{\text{Minimize } l \\
\text{ subject to } \\
\sum_{i = 1}^{I} x_i r_{ij} + y_j \geq t_j \forall j \in J \\
l \geq \frac{y_j}{t_j} \forall j \in J \\
\sum_{i = 1}^{I} x_i c_i \leq B}{
Minimize l subject to
sum_i^I (xi * rij) + yj >= tj for all j in J &
l >= (yj / tj) for all j in J &
sum_i^I (xi * ci) <= B}
}
\seealso{
Other tests:
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-long-lines}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-sexpr-title}},
\code{\link{test-tables}},
\code{\link{test-verbatim}}
}
\concept{tests}
\keyword{internal}
