% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-params}
\alias{test-params}
\title{Test case: params}
\arguments{
\item{...}{ellipsis}
}
\description{
Test case: params
}
\seealso{
Other tests:
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-long-lines}},
\code{\link{test-math-examples}},
\code{\link{test-output-styles}},
\code{\link{test-sexpr-title}},
\code{\link{test-tables}},
\code{\link{test-verbatim}}
}
\concept{tests}
\keyword{internal}
