% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-sections}
\alias{test-sections}
\title{Test case: section highlighting}
\description{
Test case: section highlighting
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{rlang::is_installed()
}\if{html}{\out{</div>}}

bla\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{url: https://pkgdown.r-lib.org
footer:
 left:
   structure: [pkgdown]
 right:
   structure: [authors, legal]
   components:
template:
  bootstrap: 4
}\if{html}{\out{</div>}}
}
\seealso{
Other tests: 
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-long-lines}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-sexpr-title}},
\code{\link{test-verbatim}}
}
\concept{tests}
\keyword{internal}
