% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-sexpr-title}
\alias{test-sexpr-title}
\title{Test case: \Sexpr[stage=render,results=rd]{"sexpr"}}
\description{
Test case: \Sexpr[stage=render,results=rd]{"sexpr"}
}
\seealso{
Other tests:
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-long-lines}},
\code{\link{test-math-examples}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-tables}},
\code{\link{test-verbatim}}
}
\concept{tests}
\keyword{internal}
