% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-verbatim}
\alias{test-verbatim}
\title{Test case: preformatted blocks & syntax highlighting}
\description{
Manual test cases for various ways of embedding code in sections.
All code blocks should have copy and paste button.
}
\section{Should be highlighted}{
Valid R code in \verb{\preformatted{}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mean(a + 1)
}\if{html}{\out{</div>}}

R code in \code{R} block:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{mean(a + 1)
}\if{html}{\out{</div>}}

R code in \code{r} block:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mean(a + 1)
}\if{html}{\out{</div>}}

Yaml

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{yaml: [a, 1]
}\if{html}{\out{</div>}}
}

\section{Shouldn't be highlighted}{
Non-R code in \verb{\preformatted{}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{yaml: [a, b, c]
}\if{html}{\out{</div>}}
}

\seealso{
Other tests:
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-long-lines}},
\code{\link{test-math-examples}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-sexpr-title}},
\code{\link{test-tables}}
}
\concept{tests}
\keyword{internal}
