% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-widgets}
\alias{test-widgets}
\title{Test case: widgets}
\description{
Test case: widgets
}
\examples{

library(leaflet)
leaflet() \%>\%
  addTiles() \%>\%
  addMarkers(
    lng=174.768,
    lat=-36.852,
    popup="The birthplace of R"
  )

library(d3heatmap)
d3heatmap(mtcars, scale = "column", colors = "Blues")

library(DT)
datatable(iris)
}
\seealso{
Other tests: \code{\link{test-dont}},
  \code{\link{test-figures}}, \code{\link{test-links}},
  \code{\link{test-lists}}, \code{\link{test-params}}
}
\keyword{internal}
