% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp.R
\name{warp}
\alias{warp}
\title{warp}
\usage{
warp(x, trans = 1, ...)
}
\arguments{
\item{x}{\code{vector} time-series or \code{data.frame}, to be
warped}

\item{trans}{\code{numeric vector} the warp transformation to apply
to \code{x}.}

\item{...}{Other arguments, currently ignored.}
}
\value{
\code{x} warped, by default a \code{data.frame}.
}
\description{
Warps a supplied \code{vector} time-series or
\code{data.frame} using a warp transformation.
}
\note{
\code{warp}s are not conservative.
}
\examples{
x <- 1:1000
plot(x, type = "l", main = "Simple Warping")
#transformations < 1 compress time-series
lines(warp(x, 0.5), col = "red")
#transformations > 1 expand time-series
#to expand and the compress, and end at original end value
#   mean(transformation) should equal 1
lines(warp(x, c(1.75, 0.25)), col = "blue")
legend(725, 350, c("0.5", "1.75, 0.25"),
       col=c("red", "blue"), lty=1, cex=0.75)
}
\seealso{
\code{\link{approx}}
}
\author{
Karl Ropkins
}
