% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{write_index_r}
\alias{write_index_r}
\title{Write index.R file for all help topics}
\usage{
write_index_r(pkg = ".")
}
\arguments{
\item{pkg}{path to source version of package.  See
\code{\link[devtools]{as.package}} for details on how paths and package
names are resolved.}
}
\description{
This function will produce a file `index-all.R` in `inst/staticdocs` that you
can edit and resave as `index.R`. This will generate a custom ordering of the
function reference section of a static docs site.
}
\seealso{
\code{\link{sd_section}}, \code{\link{write_index_yaml}}
}

