% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{write_index_yaml}
\alias{write_index_yaml}
\title{Write yaml index file for all help topics}
\usage{
write_index_yaml(pkg = ".")
}
\arguments{
\item{pkg}{path to source version of package.  See
\code{\link[devtools]{as.package}} for details on how paths and package
names are resolved.}
}
\description{
This function will produce a file `index-all.yaml` in `inst/staticdocs` that
you can edit and resave as `index.yaml`. This will then be accepted as an
alternative to the `index.R` used to generate a custom ordering of the
function reference section of a static docs site.
}
\seealso{
\code{\link{sd_section}}, \code{\link{write_index_r}}
}

