% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{as.sd_package}
\alias{as.sd_package}
\title{Return information about a package}
\usage{
as.sd_package(pkg = ".", site_path = "inst/web", examples = TRUE,
  templates_path = "inst/staticdocs/templates",
  bootstrap_path = "inst/staticdocs/bootstrap", mathjax = TRUE)
}
\arguments{
\item{pkg}{name of package, as character vector}

\item{site_path}{root directory in which to create documentation. The
default, \code{NULL}, first looks at the value of \code{site_path} set in
\file{DESCRIPTION}, and if not found uses \code{inst/staticdocs}.}

\item{examples}{include examples or not? The default, \code{NULL}, first
looks at the value of \code{examples} set in \file{DESCRIPTION}, and if
not found uses \code{TRUE}.}

\item{templates_path}{a specific directory path to use when searching for
rendering templates, in addition to the default locations of
packagedir/inst/staticdocs, packagedir/staticdocs, and the staticdocs
package's included templates directory.}

\item{bootstrap_path}{a specific directory path to use when searching for
bootstrap style files, in addition to the default locations of
packagedir/inst/staticdocs, packagedir/staticdocs, and the staticdocs
package's included bootstrap directory.}

\item{mathjax}{whether to use mathjax to render math symbols.}
}
\value{
A named list of useful metadata about a package
}
\description{
Return information about a package
}
\keyword{internal}

