% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{as.sd_package}
\alias{as.sd_package}
\title{Return information about a package}
\usage{
as.sd_package(pkg = ".", ...)
}
\arguments{
\item{pkg}{name of package, as character vector}
}
\value{
A named list of useful metadata about a package
}
\description{
Return information about a package
}
\keyword{internal}

