% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{as_pkgdown}
\alias{as_pkgdown}
\title{Generate pkgdown data structure}
\usage{
as_pkgdown(path = ".")
}
\arguments{
\item{path}{Path to package}
}
\description{
You will generally not need to use this unless you need a custom site
design and you're writing your own equivalent of \code{\link{build_site}}.
}
