% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolink_html.R
\name{autolink_html}
\alias{autolink_html}
\title{Automatically link references and articles in an HTML page}
\usage{
autolink_html(input, output = input, local_packages = character())
}
\arguments{
\item{input, output}{Input and output paths for HTML file}

\item{local_packages}{A named character vector providing relative paths
(value) to packages (name) that can be reached with relative links
from the target HTML document.}
}
\description{
Deprecated: please use \link[downlit:downlit_html_path]{downlit::downlit_html_path} instead.
}
\examples{
\dontrun{
autolink_html("path/to/file.html",
  local_packages = c(
    shiny = "shiny",
    shinydashboard = "shinydashboard"
  )
)
}
}
\keyword{internal}
