% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-articles.R
\name{build_articles}
\alias{build_articles}
\title{Build articles}
\usage{
build_articles(pkg = ".", path)
}
\arguments{
\item{pkg}{Path to package source.}

\item{path}{Output path.}
}
\description{
Build articles
}

