% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-articles.R
\name{build_articles}
\alias{build_articles}
\title{Build articles}
\usage{
build_articles(pkg = ".", path = "docs/articles", depth = 1L)
}
\arguments{
\item{pkg}{Path to source package.}

\item{path}{Output path.}

\item{depth}{Depth of path relative to root of documentation.
Used to adjust relative links in the navbar.}
}
\description{
Each Rmarkdown vignette in \code{vignettes/} and its subdirectories is
rendered. Vignettes are rendered using a special document format that
reconciles \code{\link[rmarkdown]{html_document}()} with your staticdocs
template.
}
\section{Supressing vignettes}{


If you want articles that are not vignettes, either put them in
subdirectories or list in \code{.Rbuildignore}. An articles link
will be automatically added to the default navbar if the vignettes
directory is present: if you do not want this, you will need to
customise the navbar. See \code{\link{build_site}} details.
}

