% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-articles.R
\name{build_articles}
\alias{build_articles}
\title{Build articles}
\usage{
build_articles(pkg = ".", path, depth = 0L)
}
\arguments{
\item{pkg}{Path to source package.}

\item{path}{Output path.}

\item{depth}{Depth of path relative to root of documentation.
Used to adjust relative links in the navbar.}
}
\description{
Each vignette in \code{vignettes/} and its subdirectories is rendered.
If you want articles that are not vignettes, either put them in
subdirectories or list in \code{.Rbuildignore}.
}

