% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-home.R
\name{build_home}
\alias{build_home}
\title{Build home section}
\usage{
build_home(pkg = ".", override = list(), preview = NA,
  quiet = TRUE)
}
\arguments{
\item{pkg}{Path to package.}

\item{override}{An optional named list used to temporarily override
values in \code{_pkgdown.yml}}

\item{preview}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}

\item{quiet}{Set to \code{FALSE} to display output of knitr and
pandoc. This is useful when debugging.}
}
\description{
This function generates the home pagea, for converting other \code{.md} files
typically found in the package root and in \code{.github/}. It generates the home
page from  \code{index.md} or \code{README.md}, falling back to the description field
in \code{DESCRIPTION} if neither is found. It also builds an authors page
from the \code{DESCRIPTION} and \code{inst/CITATION} (if present) and
a license page, and converts any \code{.md} files found in
}
\section{Images and figures}{

If you want to include images in your \code{README.md}, they must be stored
somewhere in the package so that they can be displayed on the CRAN website.
The best place to put them is \code{man/figures}. If you are generating figures
with R Markdown, make sure you set up \code{fig.path} as followed:

\preformatted{
```\{r, include = FALSE\}
knitr::opts_chunk$set(
  fig.path = "man/figures/"
)
```
}
}

\section{Package logo}{

If you have a package logo, you can include it at the top of your README in a
level-one heading:\preformatted{# pkgdown <img src="man/figures/logo.png" align="right" />
}

\code{\link[=init_site]{init_site()}} will also automatically create a favicon set from your package
logo.
}

\section{YAML config - home}{

To tweak the home page, you need a section called \code{home}.

The sidebar links are automatically generated by inspecting the \code{URL} and
\code{BugReports} fields of the \code{DESCRIPTION}. You can add additional links with a
subsection called \code{links}, which should contain a list of \code{text} + \code{href}
elements:\preformatted{home:
  links:
  - text: Link text
    href: http://website.com
}

You can remove the first heading with:\preformatted{home:
  strip_header: true
}
}

\section{YAML config - home title and description}{


The homepage title that will be displayed in the browser tab,
and the title of homepage in Twitter/Open Graph metadata ("og:title"),
is by default the package Title from \code{DESCRIPTION}.

The description of the homepage in Twitter/Open Graph metadata ("og:description")
is by default the package Description from \code{DESCRIPTION}, with words quoted
with '' unquoted.

You can override these values by adding a \code{home} section in the pkgdown
config file, and add \code{title} and/or \code{description}, with words quoted by
'' unquoted.\preformatted{home:
  title: This is my package title
  description: This is a free description of my package
}
}

\section{YAML config - authors}{

The "developers" list is populated by the maintainer ("cre"), authors
("aut"), and funder ("fnd") from the \code{DESCRIPTION}. You can modify their
display on the home page by adding a subsection for \code{authors}. The author
name should match the value in \code{DESCRIPTION}:\preformatted{authors:
  name:
    href: "http://name-website.com"
    html: "<img src='name-picture.png' height=24>"
}
}

\section{ORCID}{

ORCID identification numbers are added to author info if they
are defined in the \code{DESCRIPTION}:\preformatted{person("Jay", "Hesselberth", role = "aut",
  comment = c(ORCID = "0000-0002-6299-179X")
),
}
}

\section{Badges}{

Status badges are displayed in the sidebar under the section "Dev status".
This section is automatically populated if there is an identifiable badges
paragraph in the homepage source (index.Rmd, index.md, README.Rmd,
README.md):
\itemize{
\item pkgdown first looks for a paragraph starting with \code{<!-- badges: start -->}
and ending with \code{<!-- badges: end -->} as created by
\code{usethis::use_readme_md()} or \code{usethis::use_readme_rmd()};
\item failing that, pkgdown looks at the first paragraph, and uses it if it
only contains images.
}
}

