% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-index.R
\name{build_home}
\alias{build_home}
\title{Build home page}
\usage{
build_home(pkg = ".", path = NULL)
}
\arguments{
\item{path}{Output path}
}
\description{
First looks for \code{README.md}, then \code{index.md}. If neither is
found, falls back to the description field in \code{DESCRIPTION}.
}

