% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-home.R
\name{build_home}
\alias{build_home}
\title{Build home page}
\usage{
build_home(pkg = ".", path = "docs", depth = 0L)
}
\arguments{
\item{pkg}{Path to source package.}

\item{path}{Output path.}

\item{depth}{Depth of path relative to root of documentation.
Used to adjust relative links in the navbar.}
}
\description{
First looks for \code{index.Rmd} or \code{README.Rmd}, then
\code{index.md} or \code{README.md}. If none are found, falls back to the
description field in \code{DESCRIPTION}.
}
\section{YAML config}{

To tweak the home page, you need a section called \code{home}.

You can add additional links to the sidebar with a subsection called
\code{links}, which should contain a list of text + href elements:

\preformatted{
home:
  links:
  - text: Link text
    href: http://website.com
}
}

