% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-home.R
\name{build_home}
\alias{build_home}
\title{Build home page}
\usage{
build_home(pkg = ".", preview = NA)
}
\arguments{
\item{pkg}{Path to source package. If R working directory is not
set to the source directory, then pkg must be a fully qualified
path to the source directory (not a relative path).}

\item{preview}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}
}
\description{
First looks for \code{index.Rmd} or \code{README.Rmd}, then
\code{index.md} or \code{README.md}. If none are found, falls back to the
description field in \code{DESCRIPTION}.
}
\section{YAML config}{

To tweak the home page, you need a section called \code{home}.

The sidebar links are automatically generated by inspecting the
\code{URL} and \code{BugReports} fields of the \code{DESCRIPTION}.
You can add additional links with a subsection called \code{links},
which should contain a list of \code{text} + \code{href} elements:

\preformatted{
home:
  links:
  - text: Link text
    href: http://website.com
}

The "developers" list is populated by the maintainer ("cre"), authors
("aut"), and funder ("fnd").

You can remove the first heading with
\preformatted{
home:
  strip_header: true
}
}

\section{Badges}{

Status badges are displayed in the sidebar under the section "Dev status".
This section is automatically populated if the first paragraph of the
homepage consists solely of status badges as linked images.
}

