% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-home.R
\name{build_home}
\alias{build_home}
\title{Build home page}
\usage{
build_home(pkg = ".", path = "docs", depth = 0L, encoding = "UTF-8")
}
\arguments{
\item{pkg}{Path to source package. If R working directory is not
set to the source directory, then pkg must be a fully qualified
path to the source directory (not a relative path).}

\item{path}{Output path. Relative paths are taken relative to the
\code{pkg} directory.}

\item{depth}{Depth of path relative to root of documentation.  Used
to adjust relative links in the navbar.}

\item{encoding}{The encoding of the input files.}
}
\description{
First looks for \code{index.Rmd} or \code{README.Rmd}, then
\code{index.md} or \code{README.md}. If none are found, falls back to the
description field in \code{DESCRIPTION}.
}
\section{YAML config}{

To tweak the home page, you need a section called \code{home}.

The sidebar links are automatically generated by inspecting the
\code{URL} and \code{BugReports} fields of the \code{DESCRIPTION}.
You can add additional links with a subsection called \code{links},
which should contain a list of \code{text} + \code{href} elements:

\preformatted{
home:
  links:
  - text: Link text
    href: http://website.com
}

The "developers" list is populated by the maintainer ("cre"), authors
("aut"), and funder ("fnd").
}

