% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-home.R
\name{build_home}
\alias{build_home}
\title{Build home page}
\usage{
build_home(pkg = ".", path = "docs", depth = 0L, encoding = "UTF-8")
}
\arguments{
\item{pkg}{Path to source package. If R working directory is not
set to the source directory, then pkg must be a fully qualified
path to the source directory (not a relative path).}

\item{path}{Output path. Relative paths are taken relative to the
\code{pkg} directory.}

\item{depth}{Depth of path relative to root of documentation.  Used
to adjust relative links in the navbar.}

\item{encoding}{The encoding of the input files.}
}
\description{
First looks for \code{index.Rmd} or \code{README.Rmd}, then
\code{index.md} or \code{README.md}. If none are found, falls back to the
description field in \code{DESCRIPTION}.
}
\section{YAML config}{

To tweak the home page, you need a section called \code{home}.

The sidebar will default to containing five sections: links,
code license, citation rules, developers, and dev status.
You can override this ordering or suppress some sections by
adding a subsection called \code{sections}, which should contain a
list of section name elements:

\preformatted{
home:
  sections:
    - links
    - license
    - citation
    - authors
    - dev_status
}

The default links are, in order: CRAN / Bioconductor package
download links; GitHub / BitBucket repo links (generated by
inspecting the \code{URL} fields of the \code{DESCRIPTION} file); bug
report links (generated by inspecting the \code{BugReports} field
of the \code{DESCRIPTION} file); and any custom links you provide.

You can change the order of the links or suppress some
by adding a subsection called \code{link_order}, which should
consists of a list of section name elements:

\preformatted{
home:
  link_order:
    - repo
    - github
    - bug_report
    - custom
}

You can add additional links with a subsection called
\code{links}, which should contain a list of \code{text} + \code{href}
elements:

\preformatted{
home:
  links:
  - text: Link text
    href: http://website.com
}

The code license section is generated by reading the
\code{LICENSE} field of the \code{DESCRIPTION} file. Standard
licenses are automatically linked.

The citation section is generated by reading the
\code{inst/CITATION} file from the package.

The "developers" list is by default populated by the maintainer
("cre"), authors ("aut"), and funder ("fnd") but you can
override this with a section called \code{roles_include}, which
should contain a list of members:

\preformatted{
roles_include:
  - cre
  - aut
  - fnd
}

This section must be in the base level of the YML file, not
a subsection under \code{home}: this is because in addition to
populating the "developers" list, these names also populate
the page footer developer list.

Dev status is populated by pulling any badges from the home
page contents into the sidebar.
}

