% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-llm.R
\name{build_llm_docs}
\alias{build_llm_docs}
\title{Build docs for LLMs}
\usage{
build_llm_docs(pkg = ".")
}
\arguments{
\item{pkg}{Path to package.}
}
\description{
\code{build_llm_docs()} creates an \code{LLMs.txt} at the root of your site
that contains the contents of your \code{README.md}, your reference index,
and your articles index. It also creates a \code{.md} file for every existing
\code{.html} file in your site. Together, this gives an LLM an overview of your
package and the ability to find out more by following links.

If you don't want these files generated for your site, you can opt-out by
adding the following to your \code{pkgdown.yml}:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{llm-docs: false
}\if{html}{\out{</div>}}
}
\seealso{
Other site components: 
\code{\link{build_articles}()},
\code{\link{build_home}()},
\code{\link{build_news}()},
\code{\link{build_reference}()},
\code{\link{build_tutorials}()}
}
\concept{site components}
