% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-news.R
\name{build_news}
\alias{build_news}
\title{Build news section}
\usage{
build_news(pkg = ".", path = "docs/news", one_page = TRUE, depth = 1L)
}
\arguments{
\item{pkg}{Path to source package.}

\item{path}{Output path.}

\item{one_page}{If \code{TRUE}, writes all news to a single file.
If \code{FALSE}, writes one file per major version.}

\item{depth}{Depth of path relative to root of documentation.
Used to adjust relative links in the navbar.}
}
\description{
Your \code{NEWS.md} is parsed in to sections based on your use of
headings. Each minor version (i.e. the combination of first and second
components) gets on one page, with all patch versions (i.e. the third
comoponent) on a single page. News items for development versions (by
convention those versions with a four component) are displayed on an
an "unreleased" page.
}
\section{YAML config}{


There are currently no configuration options.
}

