% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-news.R
\name{build_news}
\alias{build_news}
\title{Build news section}
\usage{
build_news(pkg = ".", path = "docs/news", one_page = TRUE, depth = 1L)
}
\arguments{
\item{pkg}{Path to source package. If R working directory is not
set to the source directory, then pkg must be a fully qualified
path to the source directory (not a relative path).}

\item{path}{Output path. Relative paths are taken relative to the
\code{pkg} directory.}

\item{one_page}{If \code{TRUE}, writes all news to a single file.
If \code{FALSE}, writes one file per major version.}

\item{depth}{Depth of path relative to root of documentation.  Used
to adjust relative links in the navbar.}
}
\description{
Your \code{NEWS.md} is parsed in to sections based on your use of
headings. Each minor version (i.e. the combination of first and second
components) gets on one page, with all patch versions (i.e. the third
component) on a single page. News items for development versions (by
convention those versions with a fourth component) are displayed on an
an "unreleased" page.
}
\details{
The \code{NEWS.md} file should be formatted somewhat like this:

\preformatted{
# pkgdown 0.1.0.9000

## Major changes

- Fresh approach based on the staticdocs package. Site configuration now based on YAML files.

...
}

Commonly used subsection headers include 'Major changes', 'Bug fixes', 'Minor changes'.
}
\section{YAML config}{


There are currently no configuration options.
}

