% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-reference.R
\name{build_reference}
\alias{build_reference}
\title{Generate reference index and topics.}
\usage{
build_reference(pkg = ".", path = NULL, depth = 1L)
}
\arguments{
\item{pkg}{Path to package source.}

\item{path}{Output path.}

\item{depth}{Depth of path relative to root of documentation.
Used to adjust relative links in the navbar.}
}
\description{
Generate reference index and topics.
}

