% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-reference.R
\name{build_reference}
\alias{build_reference}
\title{Build reference section}
\usage{
build_reference(pkg = ".", examples = TRUE, run_dont_run = FALSE,
  mathjax = TRUE, seed = 1014, path = NULL, depth = 1L)
}
\arguments{
\item{pkg}{Path to package source.}

\item{examples}{Run examples?}

\item{run_dont_run}{Run examples that are surrounded in \\dontrun?}

\item{mathjax}{Use mathjax to render math symbols?}

\item{seed}{Seed used to initialize so that random examples are
reproducible.}

\item{path}{Output path.}

\item{depth}{Depth of path relative to root of documentation.
Used to adjust relative links in the navbar.}
}
\description{
Build reference section
}

