% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-reference.R
\name{build_reference}
\alias{build_reference}
\title{Generate reference index and topics.}
\usage{
build_reference(pkg = ".", path = NULL)
}
\arguments{
\item{path}{Path in which to save files}
}
\description{
Generate reference index and topics.
}

