% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-reference.R
\name{build_reference}
\alias{build_reference}
\alias{build_reference_index}
\title{Build reference section}
\usage{
build_reference(pkg = ".", examples = TRUE, run_dont_run = FALSE,
  mathjax = TRUE, seed = 1014, path = "docs/reference", depth = 1L)

build_reference_index(pkg = ".", path = "docs/reference", depth = 1L)
}
\arguments{
\item{pkg}{Path to source package.}

\item{examples}{Run examples?}

\item{run_dont_run}{Run examples that are surrounded in \\dontrun?}

\item{mathjax}{Use mathjax to render math symbols?}

\item{seed}{Seed used to initialize so that random examples are
reproducible.}

\item{path}{Output path. Relative paths are taken relative to the
\code{pkg} directory.}

\item{depth}{Depth of path relative to root of documentation.
Used to adjust relative links in the navbar.}
}
\description{
By default, pkgdown will generate an index that simply lists all
the functions in alphabetical order. To override this, provide a
\code{reference} section in your \code{_pkgdown.yml} as described
below.
}
\section{YAML config}{

To tweak the index page, you need a section called \code{reference}
which provides a list of sections containing, a \code{title}, list of
\code{contents}, and optional \code{description}.

For example, the following code breaks up the functions in pkgdown
into two groups:

\preformatted{
reference:
- title: Render components
  desc:  Build each component of the site.
  contents:
  - starts_with("build_")
  - init_site
- title: Templates
  contents:
  - render_page
}

Note that \code{contents} can contain either a list of function names,
or if the functions in a section share a common prefix or suffix, you
can use \code{starts_with("prefix")} and \code{ends_with("suffix")} to
select them all. For more complex naming schemes you can use an aribrary
regular expression with \code{matches("regexp")}.

pkgdown will check that all non-internal topics are included on
this page, and will generate a warning if you have missed any.
}

