% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_site}
\alias{build_site}
\title{Build a complete pkgdown website}
\usage{
build_site(
  pkg = ".",
  examples = TRUE,
  run_dont_run = FALSE,
  seed = 1014,
  lazy = FALSE,
  override = list(),
  preview = NA,
  devel = FALSE,
  new_process = !devel,
  install = !devel,
  document = "DEPRECATED"
)
}
\arguments{
\item{pkg}{Path to package.}

\item{examples}{Run examples?}

\item{run_dont_run}{Run examples that are surrounded in \\dontrun?}

\item{seed}{Seed used to initialize so that random examples are
reproducible.}

\item{lazy}{If \code{TRUE}, will only rebuild articles and reference pages
if the source is newer than the destination.}

\item{override}{An optional named list used to temporarily override
values in \verb{_pkgdown.yml}}

\item{preview}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}

\item{devel}{Use development or deployment process?

If \code{TRUE}, uses lighter-weight process suitable for rapid
iteration; it will run examples and vignettes in the current process,
and will load code with \code{pkgload::load_all()}.

If \code{FALSE}, will first install the package to a temporary library,
and will run all examples and vignettes in a new process.

\code{build_site()} defaults to \code{devel = FALSE} so that you get high fidelity
outputs when you building the complete site; \code{build_reference()},
\code{build_home()} and friends default to \code{devel = TRUE} so that you can
rapidly iterate during development.}

\item{new_process}{If \code{TRUE}, will run \code{build_site()} in a separate process.
This enhances reproducibility by ensuring nothing that you have loaded
in the current process affects the build process.}

\item{install}{If \code{TRUE}, will install the package in a temporary library
so it is available for vignettes.}

\item{document}{\strong{Deprecated} Use \code{devel} instead.}
}
\description{
\code{build_site()} is a convenient wrapper around six functions:
\itemize{
\item \code{\link[=init_site]{init_site()}}
\item \code{\link[=build_home]{build_home()}}
\item \code{\link[=build_reference]{build_reference()}}
\item \code{\link[=build_articles]{build_articles()}}
\item \code{\link[=build_tutorials]{build_tutorials()}}
\item \code{\link[=build_news]{build_news()}}
}

See the documentation for the each function to learn how to control
that aspect of the site.

Note if names of generated files were changed, you will need to use
\code{\link[=clean_site]{clean_site()}} first to clean up orphan files.
}
\section{YAML config}{

There are five top-level YAML settings that affect the entire site:
\code{destination}, \code{url}, \code{title}, \code{template}, and \code{navbar}.

\code{destination} controls where the site will be generated. It defaults to
\verb{docs/} (for GitHub pages), but you can override if desired. Relative
paths will be taken relative to the package root.

\code{url} is optional, but strongly recommended.
It specifies where the site will be published and is used to:
\itemize{
\item Allow other pkgdown sites to link to your site when needed,
rather than using generic links to \url{https://rdrr.io}.
See \code{vignette("linking")} for more information.
\item Generate a \code{sitemap.xml}, increasing the searchability of your site.
\item Automatically generate a \code{CNAME} when
\link[=deploy_site_github]{deploying to github}.
\item Generate metadata used by Twitter and the Open Graph protocol
for rich social media cards, see \code{vignette("metadata")}.
\item Adds the "external-link" class to external links
for sites using BS4, see \code{vignette("customise")}.
}\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{url: https://pkgdown.r-lib.org
}\if{html}{\out{</div>}}

\code{title} overrides the default site title, which is the package name.
It's used in the page title and default navbar.

You can also provided information to override the default display of
the authors. Provided a list named with the name of each author,
including \code{href} to add a link, or \code{html} to override the
text:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{authors:
  Hadley Wickham:
    href: http://hadley.nz
  RStudio:
    href: https://www.rstudio.com
    html: <img src="https://www.tidyverse.org/rstudio-logo.svg" height="24" />
}\if{html}{\out{</div>}}
}

\section{Development mode}{

The development mode of a site controls four main things:
\itemize{
\item Where the site is built.
\item The colour of the package version in the navbar.
\item The optional tooltip associated with the version.
\item The indexing of the site by search engines.
}

You can override the default development mode by adding a
new \code{development} field to \verb{_pkgdown.yml}, e.g.\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{development:
  mode: devel
}\if{html}{\out{</div>}}

There are currently four possible modes:
\itemize{
\item \strong{release} (\code{mode: release}), the default. Site is written to \verb{docs/}.
Version in navbar gets the default colouring.
\item \strong{development} (\code{mode: devel}). Site is written to \verb{docs/dev/}.
Version in navbar gets the "danger" class and a message stating these are
docs for an in-development version of the package. The \code{noindex} meta tag
is used to ensure that these packages are not indexed by search engines.
\item \strong{unreleased} (\code{mode: unreleased}). Site is written to \verb{docs/}.
Version in navbar gets the "danger" class, and a message indicating the
package is not yet on CRAN.
\item \strong{automatic} (\code{mode: auto}): pkgdown automatically detects the mode
based on the version number:
\itemize{
\item \verb{0.0.0.9000} (\verb{0.0.0.*}): unreleased.
\item four version components: development.
\item everything else -> release.
}
}

There are three other options that you can control:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{development:
  destination: dev
  version_label: danger
  version_tooltip: "Custom message here"
}\if{html}{\out{</div>}}

\code{destination} allows you to override the default subdirectory used for the
development site; it defaults to \verb{dev/}. \code{version_label} allows you to
override the style used for development (and unreleased) versions of the
package. It defaults to "danger", but you can set to "default", "info", or
"warning" instead. (The precise colours are determined by your bootstrap
theme, but become progressively more eye catching as you go from default
to danger). Finally, you can choose to override the default tooltip with
\code{version_tooltip}.
}

\section{YAML config - navbar}{

You can customise the navigation bar that appears at the top of the page with the \code{navbar} field.
It's made up of two pieces: \code{structure}, which defines the overall layout, and \code{components}, which defines what each piece looks like.
This organisation makes it easy to mix and match pkgdown defaults with your own customisations.

This is the default structure:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{navbar:
  structure:
    left:  [intro, reference, articles, tutorials, news]
    right: [search, github]
}\if{html}{\out{</div>}}

It makes use of the the six built-in components:
\itemize{
\item \code{intro}: "Get Started", which links to a vignette with the same name as the package.
\item \code{reference}, if there are any \code{.Rd} files.
\item \code{articles}, if there are any vignettes or articles.
\item \code{tutorials}, if there any tutorials.
\item \code{news}, if \code{NEWS.md} exists.
\item \code{search}, the search box (see \code{vignette("search")} for more details).
\item \code{github}, a link to the source repository (with an icon), if it can be automatically determined from the \code{DESCRIPTION}.
}

You can use the \code{structure} field to reorganise the navbar without changing the default contents:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{navbar:
  structure:
    left:  [search]
    right: [reference, articles]
}\if{html}{\out{</div>}}

You can use \code{components} to override the default content.
For example, this yaml provides a custom articles menu:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{navbar:
 components:
   articles:
    text: Articles
    menu:
    - text: Category A
    - text: Title A1
      href: articles/a1.html
    - text: Title A2
      href: articles/a2.html
    - text: -------
    - text: "Category B"
    - text: Article B1
      href: articles/b1.html
}\if{html}{\out{</div>}}

Components uses the same syntax as \href{https://bookdown.org/yihui/rmarkdown/rmarkdown-site.html#site-navigation}{RMarkdown menus}.
The elements of \code{menu} can be:
\itemize{
\item A link (\code{text} + \code{href})
\item A heading (just \code{text})
\item A separator (\verb{text: ——–})
}

Instead of text, you can also use the name of an \code{icon}s from \href{https://fontawesome.com/icons?d=gallery}{fontawesome}.
You should also provide a textual description in the \code{aria-label} field for screenreader users.

To add a new component to the navbar, you need to modify both \code{structure} and \code{components}.
For example, the following yaml adds a new "twitter" component that appears to the left of the github icon.\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{navbar:
  structure:
    right: [twitter, github]
  components:
    twitter:
      icon: fa-twitter
      href: http://twitter.com/hadleywickham
      aria-label: Twitter
}\if{html}{\out{</div>}}

You can also customise the colour scheme of the navbar by using the \code{type}
and \code{bg} parameters. See \code{vignette("customise")} for more details.
}

\section{YAML config - search}{

See \code{vignette("search")}
}

\section{YAML config - template}{

You can get complete control over the appearance of the site using the
\code{template} component. There are two components to the template:
the HTML templates used to layout each page, and the css/js assets
used to render the page in the browser.

The easiest way to tweak the default style is to use a bootswatch template,
by passing on the \code{bootswatch} template parameter to the built-in
template:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{template:
  bootswatch: cerulean
}\if{html}{\out{</div>}}

See \code{vignette("customise")} for details.

Optionally provide the \code{ganalytics} template parameter to enable
\href{https://marketingplatform.google.com/about/analytics/}{Google Analytics}.
It should correspond to your
\href{https://support.google.com/analytics/answer/1008080}{tracking id}.

When enabling Google Analytics, be aware of the type and amount of
user information that you are collecting. You may wish to limit the
extent of data collection or to add a privacy disclosure to your
site, in keeping with current laws and regulations.\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{template:
  params:
    ganalytics: UA-000000-01
}\if{html}{\out{</div>}}

Suppress indexing of your pages by web robots by setting \code{noindex: true}:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{template:
  params:
    noindex: true
}\if{html}{\out{</div>}}

You can also override the default templates and provide additional
assets. You can do so by either storing them in the
directories \code{pkgdown/assets} and \code{pkgdown/templates},
or by supplying \code{path} and \code{asset_path} pointing to alternative folders.
To suppress inclusion
of the default assets, set \code{default_assets} to false.\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{template:
  path: path/to/templates
  assets: path/to/assets
  default_assets: false
}\if{html}{\out{</div>}}

These settings are currently recommended for advanced users only. There
is little documentation, and you'll need to read the existing source
for pkgdown templates to ensure that you use the correct components.

For further information including how to provide templates and assets in
a separate package, see \code{vignette("customise")}c.

You can use the \code{trailing_slash_redirect} to automatically redirect
\code{your-package-url.com} to \verb{your-package-url.com/}, using a JS script
added to the \verb{<head>} of the home page.\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{template:
  trailing_slash_redirect: true
}\if{html}{\out{</div>}}
}

\section{YAML config - repo}{

pkgdown automatically generates links to the source repository in a few
places
\itemize{
\item Articles and documentation topics are linked back to the
underlying source file.
\item The NEWS automatically links issue numbers and user names.
\item The homepage provides a link to  "Browse source code"
}

pkgdown automatically figures out the necessary URLs if you link to a GitHub
or GitLab repo in your \code{BugReports} or \code{URL} field. Otherwise, you can
supply your own in the \code{repo} component:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{repo:
  url:
    home: https://github.com/r-lib/pkgdown/
    source: https://github.com/r-lib/pkgdown/blob/master/
    issue: https://github.com/r-lib/pkgdown/issues/
    user: https://github.com/
}\if{html}{\out{</div>}}
\itemize{
\item \code{home}: path to package home on source code repository.
\item \verb{source:}: path to source of individual file in master branch.
\item \code{issue}: path to individual issue.
\item \code{user}: path to user.
}

The varying components (e.g. path, issue number, user name) are pasted on
the end of these URLs so they should have trailing \code{/}s.

pkgdown can automatically link to Jira issues as well, but you must specify
both a custom \code{issue} URL as well as your Jira project names to auto-link in
\code{jira_projects}. You can specify as many projects as you would like in a last
(in the example below we would link both the \code{PROJ} and \code{OTHER} Jira
projects):\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{repo:
  jira_projects: [PROJ, OTHER]
  url:
    issue: https://jira.organisation.com/jira/browse/
}\if{html}{\out{</div>}}

pkgdown defaults to using the "master" branch for source file URLs. This can
be configured to use a specific branch when linking to source files by
specifying a branch name:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{repo:
  branch: main
}\if{html}{\out{</div>}}
}

\section{YAML config - deploy}{

\code{deploy} currently offers a single parameter:
\itemize{
\item \code{install_metadata} allows you to install package index metadata into
the package itself. Normally this metadata is made available on the
published site; installing it into your package means that it's
available for autolinking even if your website is not reachable at build
time (e.g. because it's only behind the firewall or requires auth).\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{deploy:
  install_metadata: true
}\if{html}{\out{</div>}}
}
}

\section{YAML config - footer}{

You can customise the top navbar with the \code{footer} field.

By default, the footer is automatically populated with:
\itemize{
\item The names of the authors on the left.
\item A link to the pkgdown website on the right.
}

You can override these defaults with the \code{footer} field.
The example below puts the authors information on the right together with a legal disclaimer, and puts pkgdown on the left.\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{footer:
  left:
    structure: [pkgdown]
  right:
    structure: [authors, legal]
    components:
      legal: Provided without ***any warranty***.
}\if{html}{\out{</div>}}

Like the navbar and sidebar, you can control both the structure and the components.
Unlike the navbar or sidebar, the components of the footer are just pasted together into a string that converted to HTML and insert in the document.

See \code{?build_home} if you want to tweak \emph{which} authors appear in the footer.
}

\section{YAML config - redirects}{

If you change the structure of your documentation (by renaming vignettes or help topics) you can setup redirects from the old content to the new content.
One or several now-absent pages can be redirected to a new page (or to a new section of a new page).
This works by creating a html page that performs a "meta refresh", which isn't the best way of doing a redirect but works everywhere that you might deploy your site.

The syntax is the following, with old paths on the left, and new paths or URLs on the right.\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{redirects:
  - ["articles/old-vignette-name.html", "articles/new-vignette-name.html"]
  - ["articles/another-old-vignette-name.html", "articles/new-vignette-name.html"]
  - ["articles/yet-another-old-vignette-name.html", "https://pkgdown.r-lib.org/dev"]
}\if{html}{\out{</div>}}

If for some reason you choose to redirect an existing page make sure to exclude it from the search index, see \code{?build_search}.
}

\section{Options}{

Users with limited internet connectivity can disable CRAN checks by setting
\code{options(pkgdown.internet = FALSE)}. This will also disable some features
from pkgdown that requires an internet connectivity. However, if it is used
to build docs for a package that requires internet connectivity in examples
or vignettes, this connection is required as this option won't apply on them.

Users can set a timeout for \code{build_site(new_process = TRUE)} with
\code{options(pkgdown.timeout = Inf)}, which is useful to prevent stalled builds from
hanging in cron jobs.
}

\examples{
\dontrun{
build_site()

build_site(override = list(destination = tempdir()))
}
}
