% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.r
\name{build_site}
\alias{build_package}
\alias{build_site}
\alias{staticdocs-package}
\title{Build complete static documentation for a package.}
\usage{
build_site(pkg = ".", site_path = "inst/web", examples = TRUE,
  templates_path = "inst/staticdocs/templates",
  bootstrap_path = "inst/staticdocs/bootstrap", mathjax = TRUE,
  with_vignettes = TRUE, with_demos = TRUE, with_readme = TRUE,
  launch = interactive())
}
\arguments{
\item{pkg}{path to source version of package.  See
\code{\link[devtools]{as.package}} for details on how paths and package
names are resolved.}

\item{site_path}{root directory in which to create documentation. The
default, \code{NULL}, first looks at the value of \code{site_path} set in
\file{DESCRIPTION}, and if not found uses \code{inst/staticdocs}.}

\item{examples}{include examples or not? The default, \code{NULL}, first
looks at the value of \code{examples} set in \file{DESCRIPTION}, and if
not found uses \code{TRUE}.}

\item{templates_path}{a specific directory path to use when searching for
rendering templates, in addition to the default locations of
packagedir/inst/staticdocs, packagedir/staticdocs, and the staticdocs
package's included templates directory.}

\item{bootstrap_path}{a specific directory path to use when searching for
bootstrap style files, in addition to the default locations of
packagedir/inst/staticdocs, packagedir/staticdocs, and the staticdocs
package's included bootstrap directory.}

\item{mathjax}{whether to use mathjax to render math symbols.}

\item{with_vignettes}{If \code{TRUE}, will build vignettes.}

\item{with_demos}{If \code{TRUE}, will build demos.}

\item{with_readme}{If \code{TRUE}, will build the README.}

\item{launch}{If \code{TRUE}, will open freshly generated site in web
browser.}
}
\description{
Currently, build_site builds documentation for:
\itemize{
  \item The package DESCRIPTION
  \item Help topics
  \item Vignettes
  \item Demos. Must be listed in \file{demos/00index}
  \item README.md files
}
}
\examples{
\dontrun{
build_site()
}
}

