% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.r
\name{build_site}
\alias{build_site}
\title{Build complete static documentation for a package.}
\usage{
build_site(pkg = ".", path = "docs", examples = TRUE,
  run_dont_run = FALSE, mathjax = TRUE, launch = interactive(),
  seed = 1014)
}
\arguments{
\item{pkg}{Path to package source.}

\item{path}{Output path.}

\item{examples}{Run examples?}

\item{run_dont_run}{Run examples that are surrounded in \\dontrun?}

\item{mathjax}{Use mathjax to render math symbols?}

\item{launch}{If \code{TRUE}, will open freshly generated site in web
browser.}

\item{seed}{Seed used to initialize so that random examples are
reproducible.}
}
\description{
\code{build_site} calls \code{\link{init_site}()},
\code{\link{build_site}()},
\code{\link{build_home}()}, and
\code{\link{build_reference}}.
See the individual documentation for how the code works.
}
\details{
By default, staticdocs will use predefined templates and assets. If you
want to override, you can set the \code{asset_path} and \code{template_path}
at the top-level of \code{staticdocs.yml}.
}
\examples{
\dontrun{
build_site()
}
}

