% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.r
\name{build_site}
\alias{build_site}
\title{Build complete static documentation for a package.}
\usage{
build_site(pkg = ".", site_path = "docs", examples = TRUE,
  run_dont_run = FALSE, templates_path = "inst/staticdocs/templates",
  assets_path = "inst/staticdocs/assets", mathjax = TRUE,
  launch = interactive(), seed = 1014)
}
\arguments{
\item{pkg}{path to source version of package.  See
\code{\link[devtools]{as.package}} for details on how paths and package
names are resolved.}

\item{site_path}{root Directory in which to create documentation.}

\item{examples}{Run examples?}

\item{run_dont_run}{Run examples that are surrounded in \\dontrun?}

\item{templates_path}{Path in which to look for templates. If this doesn't
exist will look next in \code{pkg/inst/staticdocs/templates}, then
in staticdocs itself.}

\item{assets_path}{Path in which to look for assets files. If
this doesn't exist, will use files built into staticdocs.}

\item{mathjax}{Use mathjax to render math symbols?}

\item{launch}{If \code{TRUE}, will open freshly generated site in web
browser.}

\item{seed}{Seed used to initialize so that random examples are
reproducible.}
}
\description{
\code{build_site} calls \code{\link{init_site}()},
\code{\link{build_site}()},
\code{\link{build_home}()}, and
\code{\link{build_reference}}.
See the individual documentation for how the code works.
}
\examples{
\dontrun{
build_site()
}
}

