% Generated by roxygen2 (4.0.0): do not edit by hand
\name{build_site}
\alias{build_package}
\alias{build_site}
\alias{staticdocs-package}
\title{Build complete static documentation for a package.}
\usage{
build_site(pkg = ".", ..., launch = interactive())
}
\arguments{
  \item{pkg}{path to source version of package.  See
  \code{\link[devtools]{as.package}} for details on how
  paths and package names are resolved.}

  \item{...}{Other additional arguments passed to
  \code{\link{as.sd_package}} used to override package
  defaults.}

  \item{launch}{If \code{TRUE}, will open freshly generated
  site in web browser.}
}
\description{
Currently, knitr builds documentation for:
}
\details{
\itemize{
  \item Rd files.  Files
  \item Demos. Must be listed in \file{demos/00index}.
  \item Vignettes.
}
}

