% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.r
\name{build_site}
\alias{build_package}
\alias{build_site}
\alias{staticdocs-package}
\title{Build complete static documentation for a package.}
\usage{
build_site(pkg = ".", ..., launch = interactive())
}
\arguments{
\item{pkg}{path to source version of package.  See
\code{\link[devtools]{as.package}} for details on how paths and package
names are resolved.}

\item{...}{Other additional arguments passed to \code{\link{as.sd_package}}
used to override package defaults.}

\item{launch}{If \code{TRUE}, will open freshly generated site in web
browser.}
}
\description{
Currently, build_site builds documentation for:
}
\details{
\itemize{
  \item The package DESCRIPTION
  \item Help topics
  \item Vignettes
  \item Demos. Must be listed in \file{demos/00index}
  \item README.md files
}
}

