% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.r
\name{build_site}
\alias{build_site}
\title{Build staticdocs website}
\usage{
build_site(pkg = ".", path = "docs", examples = TRUE,
  run_dont_run = FALSE, mathjax = TRUE, launch = interactive(),
  seed = 1014)
}
\arguments{
\item{pkg}{Path to source package.}

\item{path}{Output path.}

\item{examples}{Run examples?}

\item{run_dont_run}{Run examples that are surrounded in \\dontrun?}

\item{mathjax}{Use mathjax to render math symbols?}

\item{launch}{If \code{TRUE}, will open freshly generated site in web
browser.}

\item{seed}{Seed used to initialize so that random examples are
reproducible.}
}
\description{
\code{build_site()} is a convenient wrapper around three functions:
\itemize{
  \item \code{\link{build_articles}()}
  \item \code{\link{build_home}()}
  \item \code{\link{build_reference}()}
}
See the documentation for the each function to learn how to control
that aspect of the site.
}
\section{YAML config}{


There are four top-level YAML settings that affect the entire site:

\describe{
  \item{title}{Site title, used in page title and default navbar.}
  \item{bootswatch}{Customise the overall theme with a bootswatch
    theme. You can see a complete list and how they'll affect your
    site at \url{https://gallery.shinyapps.io/117-shinythemes/}.}
 \item{assets_path}{Path to directory of additional assets to be
   copied into root directory of site. Use this when you want complete
   control over the visual design of the site. Recommended for advanced
   users only.}
 \item{template_path}{Path to directory of templates. Override the
   default (the templates built-in to staticdocs) to completely
   control the display of your site. Recommended for advanced users
   only.}
}
}
\examples{
\dontrun{
build_site()
}
}

