% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.r
\name{build_site}
\alias{build_package}
\alias{build_site}
\alias{staticdocs-package}
\title{Build complete static documentation for a package.}
\usage{
build_site(pkg = ".", site_path = "docs", news_path = "docs/news",
  news_singlepage = TRUE, examples = TRUE, run_dont_run = FALSE,
  templates_path = "inst/staticdocs/templates",
  bootstrap_path = "inst/staticdocs/bootstrap", mathjax = TRUE,
  with_vignettes = TRUE, with_demos = TRUE, launch = interactive(),
  seed = 1014)
}
\arguments{
\item{pkg}{path to source version of package.  See
\code{\link[devtools]{as.package}} for details on how paths and package
names are resolved.}

\item{site_path}{root Directory in which to create documentation.}

\item{news_path}{Directory in which to create NEWS entries.}

\item{news_singlepage}{Should all NEWS be rendered in a single page?}

\item{examples}{Run examples?}

\item{run_dont_run}{Run examples that are surrounded in \\dontrun?}

\item{templates_path}{Path in which to look for templates. If this doesn't
exist will look next in \code{pkg/inst/staticdocs/templates}, then
in staticdocs itself.}

\item{bootstrap_path}{Path in which to look for bootstrap files. If
this doesn't exist, will use files built into staticdocs.}

\item{mathjax}{Use mathjax to render math symbols?}

\item{with_vignettes}{If \code{TRUE}, will build vignettes.}

\item{with_demos}{If \code{TRUE}, will build demos.}

\item{launch}{If \code{TRUE}, will open freshly generated site in web
browser.}

\item{seed}{Seed used to initialize so that random examples are
reproducible.}
}
\description{
Currently, \code{build_site} builds documentation for help topics,
vignettes, demos, and a \code{README.md}, if present.
}
\section{Home page}{


The home page is generated from \code{inst/staticdocs/README.md},
\code{README.md}, or the \code{DESCRIPTION} if neither readme file is
present. On the homepage, you should show how to install the package,
describe what it does, and provide an example of how to use it.
}
\examples{
\dontrun{
build_site()
}
}

