% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-github.R
\name{build_site_github_pages}
\alias{build_site_github_pages}
\title{Build site for GitHub pages}
\usage{
build_site_github_pages(pkg = ".", ..., dest_dir = "docs", clean = TRUE)
}
\arguments{
\item{pkg}{Path to package.}

\item{...}{Additional arguments passed to \code{\link[=build_site]{build_site()}}.}

\item{dest_dir}{Directory to build site in.}

\item{clean}{Clean all files from old site.}
}
\description{
Designed to be run as part of automated workflows for deploying
to GitHub pages. It cleans out the old site, builds the site into \code{dest_dir}
(assuming the package has already been installed), adds a \code{.nojekyll}
file to suppress rendering by Jekyll, and adds a \code{CNAME} file if needed.
}
