% Generated by roxygen2 (4.0.0): do not edit by hand
\name{build_topics}
\alias{build_topics}
\title{Generate all topic pages for a package.}
\usage{
build_topics(package)
}
\arguments{
  \item{package_info}{A list containing information about
  the package, as generated by \code{\link{package_info}}}
}
\description{
Generate all topic pages for a package.
}
\keyword{internal}

