\name{build_topics}
\alias{build_topics}
\title{Generate all topic pages for a package.}
\usage{
  build_topics(package)
}
\arguments{
  \item{package_info}{A list containing information about
  the package, as generated by \code{\link{package_info}}}

  \item{package}{path to source version of package.  See
  \code{\link[devtools]{as.package}} for details on how
  paths and package names are resolved.}
}
\description{
  Generate all topic pages for a package.
}
\keyword{internal}

