% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_pkgdown}
\alias{check_pkgdown}
\title{Check \verb{_pkgdown.yml}}
\usage{
check_pkgdown(pkg = ".")
}
\arguments{
\item{pkg}{Path to package.}
}
\description{
Check that your \verb{_pkgdown.yml} is valid without building the whole
site. Currently this:
\itemize{
\item Checks the reference and article indexes to ensure that pkgdown can
read them, and that every documentation topic and vignette/article is
included in the index.
\item Validates any opengraph metadata that you might have supplied
}
}
