% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy-site.R
\name{deploy_to_branch}
\alias{deploy_to_branch}
\title{Build and deploy a site locally}
\usage{
deploy_to_branch(
  pkg = ".",
  commit_message = construct_commit_message(pkg),
  branch = "gh-pages",
  remote = "origin",
  build_cname = (branch == "gh-pages"),
  ...
)
}
\arguments{
\item{pkg}{Path to package.}

\item{commit_message}{The commit message to be used for the commit.}

\item{branch}{The git branch to deploy to}

\item{remote}{The git remote to deploy to}

\item{build_cname}{Add a \code{CNAME} file containing the URL to the site.
Needed for GitHub pages with custom domains.}

\item{...}{Additional arguments passed to \code{\link[=build_site]{build_site()}}.}
}
\description{
Assumes that you're in a git clone of the project, and the package is
already installed.
}
