% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{fig_settings}
\alias{fig_settings}
\title{Get current settings for figures}
\usage{
fig_settings()
}
\value{
A list containing the entries from the \code{figures} field in \verb{_pkgdown.yaml}
(see \code{\link[=build_reference]{build_reference()}}), with default values added. Computed \code{width} and
\code{height} values (in pixels) are also included.
}
\description{
You will generally not need to use this function unless you are handling
custom plot output.

Packages needing custom parameters should ask users to place them within
the \code{other.parameters} entry under the package name, e.g.

\if{html}{\out{<div class="sourceCode">}}\preformatted{figures:
  other.parameters:
    rgl:
      fig.asp: 1
}\if{html}{\out{</div>}}
}
\keyword{internal}
