% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{in_pkgdown}
\alias{in_pkgdown}
\title{Determine if code is executed by pkgdown}
\usage{
in_pkgdown()
}
\description{
This is occassionally useful when you need different behaviour by
pkgdown and regular documentation.
}
\examples{
in_pkgdown()
}
