% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_site}
\alias{init_site}
\title{Initialise site infrastructure}
\usage{
init_site(pkg = ".")
}
\arguments{
\item{pkg}{Path to package.}
}
\description{
This creates the output directory (\code{docs/}), \code{favicon.ico} (from the package
logo), a machine readable description of the site, and copies CSS/JS
assets and extra files.
}
\section{Custom CSS/JS}{

If you want to do minor customisation of your pkgdown site, the easiest
way is to add \code{pkgdown/extra.css} and \code{pkgdown/extra.js}. These
will be automatically copied to \code{docs/} and inserted into the
\code{<HEAD>} after the default pkgdown CSS and JS.
}

\section{Favicon}{

If you include you package logo in the standard location of
\code{man/figures/logo.png}, a favicon will be automatically created for
you.
}

