% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.r
\name{init_site}
\alias{init_site}
\title{Initialise the site}
\usage{
init_site(path, assets_path = NULL)
}
\arguments{
\item{assets_path}{Path in which to look for assets files. If
this doesn't exist, will use files built into staticdocs.}
}
\description{
Creates directory and copies assets.
}

