\name{package_info}
\alias{package_info}
\title{Return information about a package}
\usage{
  package_info(package, base_path = NULL, examples = NULL)
}
\arguments{
  \item{package}{name of package, as character vector}
}
\value{
  A named list of useful metadata about a package
}
\description{
  Return information about a package
}
\keyword{internal}

