% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-package.R
\docType{package}
\name{pkgdown-package}
\alias{pkgdown}
\alias{pkgdown-package}
\title{pkgdown: Make Static HTML Documentation for a Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate an attractive and useful website from a source package. 'pkgdown' converts your documentation, vignettes, 'README', and more to 'HTML' making it easy to share information about your package online.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkgdown.r-lib.org/}
  \item \url{https://github.com/r-lib/pkgdown}
  \item Report bugs at \url{https://github.com/r-lib/pkgdown/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})

Authors:
\itemize{
  \item Jay Hesselberth (\href{https://orcid.org/0000-0002-6299-179X}{ORCID})
  \item Maëlle Salmon (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
  \item Olivier Roy
  \item Salim Brüggemann (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
