% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown_print.R
\name{pkgdown_print}
\alias{pkgdown_print}
\title{Print object in pkgdown output}
\usage{
pkgdown_print(x, visible)
}
\arguments{
\item{x}{Object to display}

\item{visible}{Whether it is visible or not}
}
\description{
This lets package authors control how their objects appear specifically
in pkgdown examples. The default method calls \code{\link[=print]{print()}}.
}
\keyword{internal}
