% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_page}
\alias{preview_page}
\title{Preview a local pkgdown page in the browser}
\usage{
preview_page(path, pkg = ".")
}
\description{
Only works when rendering the working directory, as this is the most
common interactive workflow.
}
\keyword{internal}
