% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_site}
\alias{preview_site}
\title{Open site in browser}
\usage{
preview_site(pkg = ".", path = ".", preview = TRUE)
}
\arguments{
\item{pkg}{Path to package.}

\item{path}{Path relative to destination}

\item{preview}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}
}
\description{
\code{preview_site()} opens your pkgdown site in your browser. pkgdown has been
carefully designed to work even when served from the file system like
this; the only part that doesn't work is search. You can use \code{servr::httw("docs/")}
to create a server to make search work locally.
}
