% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_site}
\alias{preview_site}
\title{Open site in browser}
\usage{
preview_site(pkg = ".", path = ".", preview = TRUE)
}
\arguments{
\item{pkg}{Path to package.}

\item{path}{Path relative to destination}

\item{preview}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}
}
\description{
\code{preview_site()} opens your pkgdown site in your browser, served via a
local HTTP server. This enables dynamic features such as search to work
correctly in preview.
}
\seealso{
\code{\link[=stop_preview]{stop_preview()}} to stop the server.
}
