% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_page}
\alias{render_page}
\title{Render complete page.}
\usage{
render_page(package, name, data, path = "")
}
\arguments{
\item{package}{Path to package to document.}

\item{name}{Name of the template (e.g. index, demo, topic)}

\item{data}{Data for the template}

\item{path}{Location to create file. If \code{""} (the default),
prints to standard out.}
}
\description{
Render complete page.
}

