\name{render_page}
\alias{render_page}
\title{Render complete page.}
\usage{
  render_page(package, name, data, path = "")
}
\arguments{
  \item{name}{of the template (e.g. index, demo, topic)}

  \item{data}{data for the template}

  \item{path}{location to create file.  If \code{""} (the
  default), prints to standard out.}
}
\description{
  Render complete page.
}

