% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{stop_preview}
\alias{stop_preview}
\title{Stop HTTP preview}
\usage{
stop_preview()
}
\description{
Stops the HTTP server started by \code{\link[=preview_site]{preview_site()}}, if active. This can be
called manually, but is not strictly necessary as the server is
automatically stopped when previewing a new site or ending the R session.
}
\keyword{internal}
