% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-crayon}
\alias{test-crayon}
\title{Test case: crayon}
\description{
Test case: crayon
}
\examples{
cat(crayon::red("This is red"), "\n")
cat(crayon::blue("This is blue"), "\n")

message(crayon::green("This is green"))

warning(crayon::bold("This is bold"))
}
\seealso{
Other tests: 
\code{\link{index}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-sexpr-title}},
\code{\link{test-verbatim}}
}
\concept{tests}
\keyword{internal}
