% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-dont}
\alias{test-dont}
\title{Test case: don't}
\description{
Test case: don't
}
\examples{
\dontrun{
1 + 3
}

\donttest{
1 + 3
}

answer <- 1
\dontshow{
answer <- 42
}
answer # should be 42
}
\seealso{
Other tests: \code{\link{test-figures}},
  \code{\link{test-links}}, \code{\link{test-lists}}
}
\keyword{internal}
