% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-links}
\alias{test-links}
\title{Test case: links}
\description{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{jsonlite::minify("\{\}")
#> \{\}
}\if{html}{\out{</div>}}
}
\examples{
jsonlite::minify("{}")

library(jsonlite, warn.conflicts = FALSE)
minify("{}")
}
\seealso{
Other tests: 
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-lists}},
\code{\link{test-long-lines}},
\code{\link{test-math-examples}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-sexpr-title}},
\code{\link{test-tables}},
\code{\link{test-verbatim}}
}
\concept{tests}
\keyword{internal}
