% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-links}
\alias{test-links}
\title{Test case: links}
\description{
Test case: links
}
\examples{
magrittr::subtract(10, 1)

library(magrittr, warn.conflicts = FALSE)
subtract(10, 1)
}
\seealso{
Other tests: \code{\link{test-dont}},
  \code{\link{test-figures}}, \code{\link{test-lists}},
  \code{\link{test-params}}
}
\keyword{internal}
