% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-links}
\alias{test-links}
\title{Test case: links}
\description{
\if{html}{\out{<div class="r">}}\preformatted{magrittr::subtract(10, 1)
}\if{html}{\out{</div>}}\preformatted{## [1] 9
}
}
\examples{
magrittr::subtract(10, 1)

library(magrittr, warn.conflicts = FALSE)
subtract(10, 1)
}
\seealso{
Other tests: 
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-lists}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-verbatim}}
}
\concept{tests}
\keyword{internal}
