% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-lists}
\alias{test-lists}
\title{Test cases: lists}
\description{
\itemize{\item a}

\itemize{
  \item a
  \item This is an item...

    That spans multiple paragraphs.
}

\enumerate{
  \item a
  \item b
}

\describe{
  \item{a}{1}
  \item{b}{2}
}
}
\keyword{internal}
