% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-params}
\alias{test-params}
\title{Test case: params}
\arguments{
\item{...}{ellipsis}
}
\description{
Test case: params
}
\seealso{
Other tests: \code{\link{test-dont}},
  \code{\link{test-figures}}, \code{\link{test-links}},
  \code{\link{test-lists}}
}
\keyword{internal}
