% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-verbatim}
\alias{test-verbatim}
\title{Test case: preformatted blocks & syntax highlighting}
\description{
This description block is required so that verbatim blocks are recognized
and rendered correctly.
}
\details{
\preformatted{foo

bar
}\preformatted{yaml:
  this

OR:

yaml:
  that
}

And this block should get syntax highlighting:\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{yaml:
  this

OR:

yaml:
  that
}\if{html}{\out{</div>}}
}
\seealso{
Other tests: 
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-long-lines}},
\code{\link{test-output-styles}},
\code{\link{test-params}},
\code{\link{test-sexpr-title}}
}
\concept{tests}
\keyword{internal}
