% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test-verbatim}
\alias{test-verbatim}
\title{Test case: verbatim blocks}
\description{
This description block is required so that verbatim blocks are recognized
and rendered correctly.
}
\details{
\preformatted{foo

bar
}\preformatted{yaml:
  this

OR:

yaml:
  that
}
}
\seealso{
Other tests: 
\code{\link{index}},
\code{\link{test-crayon}},
\code{\link{test-dont}},
\code{\link{test-figures}},
\code{\link{test-links}},
\code{\link{test-lists}},
\code{\link{test-output-styles}},
\code{\link{test-params}}
}
\concept{tests}
\keyword{internal}
